backend/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
fastapi==0.115.4
filelock==3.16.1
fsspec==2024.10.0
h11==0.14.0
huggingface-hub==0.26.2
idna==3.10
Jinja2==3.1.4
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.1.3
packaging==24.1
pillow==11.0.0
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
safetensors==0.4.5
setuptools==75.3.0
sniffio==1.3.1
starlette==0.41.2
sympy==1.13.1
tokenizers==0.20.2
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tqdm==4.66.6
transformers==4.46.1
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
